Move notebook-specific plots out of pydvl, rearrange some code
