+++ src/pydvl/utils/__init__.py
@@ -3,7 +3,6 @@ from .config import *
 from .dataset import *
 from .numeric import *
 from .parallel import *
-from .plotting import *
 from .progress import *
 from .types import *
 from .utility import *
